Revert "Bugfix/22 issue of completely transparent image reading"
