+++ src/clipboard/mod.rs
@@ -1,10 +1,11 @@
+#[cfg(any(target_os = "macos", target_os = "linux"))]
 pub mod arboard_clipboard;
 pub mod factory;
 pub mod traits;
 #[cfg(windows)]
 pub mod win_clipboard;
 
-#[allow(unused_imports)]
+#[cfg(any(target_os = "macos", target_os = "linux"))]
 pub use arboard_clipboard::ArboardClipboard;
 pub use factory::create_clipboard;
 pub use traits::ClipboardOperations;
