+++ tests/local_clipboard_hanlder_tests.rs
@@ -1,4 +1,3 @@
-use arboard::Clipboard;
 use bytes::Bytes;
 use chrono::Utc;
 use image::ImageReader;
@@ -7,6 +6,7 @@ use std::fs;
 use std::io::Write;
 use std::{fs::File, path::PathBuf};
 use uniclipboard::{LocalClipboardHandler, Payload};
+use arboard::Clipboard;
 
 #[tokio::test]
 #[cfg_attr(not(feature = "clipboard_tests"), ignore)]
@@ -109,82 +109,4 @@ async fn test_read_write_clipboard_image() -> Result<(), Box<dyn std::error::Err
     let local_handler = LocalClipboardHandler::new();
     local_handler.write(payload).await?;
     Ok(())
-}
-
-use image::{ImageBuffer, Rgba};
-
-#[test]
-#[cfg_attr(not(feature = "clipboard_tests"), ignore)]
-fn test_read_image_directly_from_arboard() {
-    let mut clipboard = Clipboard::new().expect("Failed to create clipboard");
-
-    // 尝试从剪贴板读取图片
-    match clipboard.get_image() {
-        Ok(image_data) => {
-            println!("Successfully read image from clipboard");
-            println!(
-                "Image dimensions: {}x{}",
-                image_data.width, image_data.height
-            );
-
-            // 将图片数据转换为 ImageBuffer
-            let img = ImageBuffer::from_raw(
-                image_data.width as u32,
-                image_data.height as u32,
-                image_data.bytes.to_vec(),
-            )
-            .expect("Failed to create image buffer");
-
-            // 保存图片以便查看
-            img.save("test_arboard_image.png")
-                .expect("Failed to save image");
-
-            // 检查图片是否有透明度
-            let has_transparency = img.pixels().any(|p: &Rgba<u8>| p[3] < 255);
-            println!("Image has transparency: {}", has_transparency);
-
-            assert!(true, "Image read successfully");
-        }
-        Err(e) => {
-            panic!("Failed to read image from clipboard: {:?}", e);
-        }
-    }
-}
-
-#[cfg(windows)]
-use clipboard_win::{formats, get_clipboard, Clipboard as WinClipboard};
-
-#[test]
-#[cfg(windows)]
-#[cfg_attr(not(feature = "clipboard_tests"), ignore)]
-fn test_read_image_using_clipboard_win() {
-    // 打开剪贴板
-    let _clipboard = WinClipboard::new().expect("Failed to open clipboard");
-
-    // 尝试从剪贴板读取位图
-    match get_clipboard(formats::Bitmap) {
-        Ok(data) => {
-            println!("Successfully read image from clipboard");
-
-            // 直接从剪贴板数据创建图像
-            let img =
-                image::load_from_memory(&data).expect("Failed to load image from clipboard data");
-
-            println!("Image dimensions: {}x{}", img.width(), img.height());
-
-            // 保存图片以便查看
-            img.save("test_clipboard_win_image.png")
-                .expect("Failed to save image");
-
-            // 检查图片是否有透明度
-            let rgba_image = img.to_rgba8();
-            let has_transparency = rgba_image.pixels().any(|p: &Rgba<u8>| p[3] < 255);
-            println!("Image has transparency: {}", has_transparency);
-
-            assert!(true, "Image read successfully");
-        }
-        Err(e) => {
-            panic!("Failed to read image from clipboard: {:?}", e);
-        }
-    }
 }
